move BMPFileIO functions
move BMPFileIO functions
